+++ MrukSample/app/src/main/java/com/meta/spatial/samples/mruksample/anchor_mesh/MrukAnchorMeshSampleActivity.kt
@@ -10,11 +10,15 @@ package com.meta.spatial.samples.mruksample.anchor_mesh
 import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
+import android.graphics.Color
+import android.net.Uri
 import android.os.Bundle
 import android.os.Environment
 import android.util.Log
+import android.view.View
 import android.widget.ArrayAdapter
 import android.widget.Button
+import android.widget.CheckBox
 import android.widget.Spinner
 import android.widget.TextView
 import com.meta.spatial.core.Entity
@@ -29,10 +33,15 @@ import com.meta.spatial.mruk.MRUKLabel
 import com.meta.spatial.mruk.MRUKLoadDeviceResult
 import com.meta.spatial.mruk.MRUKRoom
 import com.meta.spatial.mruk.MRUKSceneEventListener
+import com.meta.spatial.mruk.SceneModel
 import com.meta.spatial.physics.PhysicsFeature
 import com.meta.spatial.runtime.LayerConfig
+import com.meta.spatial.runtime.SceneMaterial
+import com.meta.spatial.runtime.SceneMesh
+import com.meta.spatial.runtime.SceneTexture
 import com.meta.spatial.samples.mruksample.MrukSampleStartMenuActivity
 import com.meta.spatial.samples.mruksample.R
+import com.meta.spatial.samples.mruksample.RoomMeshFace
 import com.meta.spatial.samples.mruksample.common.MrukInputSystem
 import com.meta.spatial.samples.mruksample.common.UIPositionSystem
 import com.meta.spatial.samples.mruksample.common.UpdateRoomSystem
@@ -42,8 +51,12 @@ import com.meta.spatial.samples.mruksample.common.returnTo2DActivity
 import com.meta.spatial.toolkit.AppSystemActivity
 import com.meta.spatial.toolkit.Grabbable
 import com.meta.spatial.toolkit.Material
+import com.meta.spatial.toolkit.Mesh
+import com.meta.spatial.toolkit.MeshCreationSystem
+import com.meta.spatial.toolkit.MeshManager
 import com.meta.spatial.toolkit.PanelRegistration
 import com.meta.spatial.toolkit.Transform
+import com.meta.spatial.toolkit.TransformParent
 import com.meta.spatial.toolkit.Visible
 import com.meta.spatial.toolkit.createPanelEntity
 import com.meta.spatial.vr.LocomotionSystem
@@ -53,13 +66,13 @@ import java.io.FileOutputStream
 import java.io.IOException
 import java.util.UUID
 
-// default activity
 class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener {
   private lateinit var mrukFeature: MRUKFeature
   private lateinit var meshSpawner: AnchorMeshSpawner
   private lateinit var procMeshSpawner: AnchorProceduralMesh
   private var globalMeshSpawner: AnchorProceduralMesh? = null
   private var currentRoomTextView: TextView? = null
+  private var checkboxRoomMesh: CheckBox? = null
 
   private var sceneDataLoaded = false
   private var showUiPanel = false
@@ -67,16 +80,20 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
 
   private val panelId = 2
 
+  private val roomMeshEntities = mutableListOf<Entity>()
+
   override fun registerFeatures(): List<SpatialFeature> {
     // Register the features that are needed for that sample. Note that in addition to the
-    // MRUKFeature the PhysicsFeature gets enabled as well. This is needed for having the physics
-    // colliders on the AnchorProceduralMesh working.
+    // MRUKFeature the PhysicsFeature gets enabled as well. This is needed for having the
+    // physics colliders on the AnchorProceduralMesh working.
     mrukFeature = MRUKFeature(this, systemManager)
     return listOf(VRFeature(this), PhysicsFeature(spatial), mrukFeature)
   }
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
+
+    componentManager.registerComponent<RoomMeshFace>(RoomMeshFace)
 
     systemManager.findSystem<LocomotionSystem>().enableLocomotion(false)
     systemManager.registerSystem(UIPositionSystem(::setUIPanelVisibility))
@@ -85,8 +102,8 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
 
     scene.enablePassthrough(true)
 
-    // Setup the AnchorMeshSpawner. The AnchorMeshSpawner is able to spawn glb/gltf meshes in place
-    // of scene anchors.
+    // Setup the AnchorMeshSpawner. The AnchorMeshSpawner is able to spawn glb/gltf meshes in
+    // place of scene anchors.
     // See the documentation AnchorMeshSpawner for more information.
 
     meshSpawner =
@@ -121,8 +138,8 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
             ),
         )
 
-    // Setup the AnchorProceduralMesh. The AnchorProceduralMesh will generate meshes in place of
-    // scene anchors. For example floor, ceiling, and walls.
+    // Setup the AnchorProceduralMesh. The AnchorProceduralMesh will generate meshes in place
+    // of scene anchors. For example floor, ceiling, and walls.
     // For more information see the documentation on AnchorProceduralMesh.
 
     val floorMaterial =
@@ -142,11 +159,16 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
                 MRUKLabel.BED to AnchorProceduralMeshConfig(null, true),
                 MRUKLabel.FLOOR to AnchorProceduralMeshConfig(floorMaterial, true),
                 MRUKLabel.WALL_FACE to AnchorProceduralMeshConfig(wallMaterial, true),
+                MRUKLabel.INVISIBLE_WALL_FACE to AnchorProceduralMeshConfig(wallMaterial, true),
+                MRUKLabel.INNER_WALL_FACE to AnchorProceduralMeshConfig(wallMaterial, true),
                 MRUKLabel.CEILING to AnchorProceduralMeshConfig(wallMaterial, true),
             ),
         )
 
     mrukFeature.addSceneEventListener(this)
+
+    val meshManager = systemManager.findSystem<MeshCreationSystem>().meshManager
+    registerRoomMeshCreator(meshManager)
 
     // Request the user to give the app scene permission if not already granted.
     // Otherwise, the app will not be able to access the device scene data.
@@ -161,40 +183,49 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
       requestPermissions(arrayOf(PERMISSION_USE_SCENE), REQUEST_CODE_PERMISSION_USE_SCENE)
     } else {
 
-      // Load the scene data from the device. If scene data was found and loaded successfully the
-      // AnchorMeshSpawner and AnchorProceduralMesh will populate the room with virtual objects.
+      // Load the scene data from the device. If scene data was found and loaded successfully
+      // the AnchorMeshSpawner and AnchorProceduralMesh will populate the room with virtual
+      // objects.
 
       loadScene(true)
     }
   }
 
-  // Implement listeners that get called whenever rooms and anchors get updated, added or removed.
-  // Hooking into these events can be useful to keep your app in sync with the loaded anchors.
+  // Implement listeners that get called whenever rooms and anchors get updated, added or
+  // removed. Hooking into these events can be useful to keep your app in sync with the loaded
+  // anchors.
   override fun onRoomAdded(room: MRUKRoom) {
-    Log.d("MRUK", "Activity: Room added: ${room.anchor}")
+    Log.d(TAG, "Activity: Room added: ${room.anchor}")
+    checkboxRoomMesh?.let {
+      if (room.roomMesh != null) {
+        it.visibility = View.VISIBLE
+      } else {
+        it.visibility = View.GONE
+      }
+    }
   }
 
   override fun onRoomRemoved(room: MRUKRoom) {
-    Log.d("MRUK", "Activity: Room removed: ${room.anchor}")
+    Log.d(TAG, "Activity: Room removed: ${room.anchor}")
   }
 
   override fun onRoomUpdated(room: MRUKRoom) {
-    Log.d("MRUK", "Activity: Room updated: ${room.anchor}")
+    Log.d(TAG, "Activity: Room updated: ${room.anchor}")
   }
 
   override fun onAnchorAdded(room: MRUKRoom, anchor: Entity) {
     val anchorUuid = anchor.getComponent<MRUKAnchor>().uuid
-    Log.d("MRUK", "Activity: Anchor $anchorUuid added to room ${room.anchor}")
+    Log.d(TAG, "Activity: Anchor $anchorUuid added to room ${room.anchor}")
   }
 
   override fun onAnchorRemoved(room: MRUKRoom, anchor: Entity) {
     val anchorUuid = anchor.getComponent<MRUKAnchor>().uuid
-    Log.d("MRUK", "Activity: Anchor $anchorUuid removed from room ${room.anchor}")
+    Log.d(TAG, "Activity: Anchor $anchorUuid removed from room ${room.anchor}")
   }
 
   override fun onAnchorUpdated(room: MRUKRoom, anchor: Entity) {
     val anchorUuid = anchor.getComponent<MRUKAnchor>().uuid
-    Log.d("MRUK", "Activity: Anchor $anchorUuid updated to room ${room.anchor}")
+    Log.d(TAG, "Activity: Anchor $anchorUuid updated to room ${room.anchor}")
   }
 
   override fun onRequestPermissionsResult(
@@ -245,10 +276,10 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
   }
 
   private fun loadScene(scenePermissionsGranted: Boolean) {
-    // MRUK has support for loading scene data from JSON files in addition to loading the scene data
-    // from device.
-    // We fallback here to JSON rooms in case the user hasn't given any scene permission just to
-    // show something.
+    // MRUK has support for loading scene data from JSON files in addition to loading the scene
+    // data from device.
+    // We fallback here to JSON rooms in case the user hasn't given any scene permission just
+    // to show something.
     // JSON rooms can be useful to accelerate development or testing different room layouts.
 
     if (!sceneDataLoaded) {
@@ -281,8 +312,8 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
     }
   }
 
-  override fun onRecenter() {
-    super.onRecenter()
+  override fun onRecenter(isUserInitiated: Boolean) {
+    super.onRecenter(isUserInitiated)
     recenterElementInView(getHmd(systemManager), Entity(panelId))
   }
 
@@ -347,6 +378,37 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
     return Environment.getExternalStorageState() == Environment.MEDIA_MOUNTED
   }
 
+  private fun showRoomMesh() {
+    hideRoomMesh()
+
+    for (room in mrukFeature.rooms) {
+      room.entity?.let { roomEntity ->
+        if (room.roomMesh != null) {
+          val roomMesh = requireNotNull(room.roomMesh)
+          for (roomMeshFace in roomMesh.faces) {
+            val roomMeshEntity =
+                Entity.create(
+                    listOf(
+                        Mesh(Uri.parse(ROOM_MESH_ID)),
+                        RoomMeshFace(room.anchor.uuid, roomMeshFace.uuid, roomMeshFace.label),
+                        Transform(),
+                        TransformParent(roomEntity),
+                    )
+                )
+            roomMeshEntities.add(roomMeshEntity)
+          }
+        }
+      }
+    }
+  }
+
+  private fun hideRoomMesh() {
+    for (roomMeshEntity in roomMeshEntities) {
+      roomMeshEntity.destroy()
+    }
+    roomMeshEntities.clear()
+  }
+
   override fun registerPanels(): List<PanelRegistration> {
     return listOf(
         PanelRegistration(R.layout.ui_anchor_mesh_menu) {
@@ -362,6 +424,27 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
             requireNotNull(rootView)
 
             currentRoomTextView = rootView?.findViewById<TextView>(R.id.mruk_current_room)
+
+            checkboxRoomMesh = rootView?.findViewById<CheckBox>(R.id.checkbox_room_mesh)
+            checkboxRoomMesh?.setOnCheckedChangeListener { _, isChecked ->
+              if (isChecked) {
+                showRoomMesh()
+              } else {
+                hideRoomMesh()
+              }
+            }
+
+            val sceneModelSpinner =
+                requireNotNull(rootView?.findViewById<Spinner>(R.id.scene_model_spinner))
+            ArrayAdapter.createFromResource(
+                    rootView?.context!!,
+                    R.array.scene_models_array,
+                    android.R.layout.simple_spinner_item,
+                )
+                .also { adapter ->
+                  adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+                  sceneModelSpinner.adapter = adapter
+                }
 
             val jsonFileSpinner =
                 requireNotNull(rootView?.findViewById<Spinner>(R.id.json_file_spinner))
@@ -381,7 +464,11 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
               jsonFileSpinner.selectedItem?.let { item ->
                 val file = applicationContext.assets.open("${item}.json")
                 val text = file.bufferedReader().use { reader -> reader.readText() }
-                mrukFeature.loadSceneFromJsonString(text)
+                mrukFeature.loadSceneFromJsonString(
+                    text,
+                    true,
+                    getSelectedSceneModel(sceneModelSpinner),
+                )
               }
             }
 
@@ -390,7 +477,9 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
 
             val loadSceneFromDeviceButton =
                 rootView?.findViewById<Button>(R.id.load_scene_from_device)
-            loadSceneFromDeviceButton?.setOnClickListener { mrukFeature.loadSceneFromDevice() }
+            loadSceneFromDeviceButton?.setOnClickListener {
+              mrukFeature.loadSceneFromDevice(true, true, getSelectedSceneModel(sceneModelSpinner))
+            }
 
             val showGlobalMeshButton = rootView?.findViewById<Button>(R.id.show_global_mesh)
             showGlobalMeshButton?.setOnClickListener { toggleGlobalMesh() }
@@ -434,10 +523,74 @@ class MrukAnchorMeshSampleActivity : AppSystemActivity(), MRUKSceneEventListener
     )
   }
 
+  private fun registerRoomMeshCreator(meshManager: MeshManager) {
+    meshManager.meshCreators.put(ROOM_MESH_ID) { entity ->
+      // Access the room face mesh component
+      val roomFaceComponent = entity.getComponent<RoomMeshFace>()
+      // Find the room mesh
+      val room = requireNotNull(mrukFeature.findRoom(roomFaceComponent.roomUuid))
+      val roomMesh = requireNotNull(room.roomMesh)
+      // Find the room face mesh
+      val roomFace = requireNotNull(roomMesh.faces.find { it.uuid == roomFaceComponent.uuid })
+      // Create the positions and index buffer for the mesh
+      val positions = FloatArray(roomFace.indices.size * 3)
+      val indices = IntArray(roomFace.indices.size)
+      roomFace.indices.forEachIndexed { i, index ->
+        positions[i * 3 + 0] = roomMesh.positions[index * 3 + 0]
+        positions[i * 3 + 1] = roomMesh.positions[index * 3 + 1]
+        positions[i * 3 + 2] = roomMesh.positions[index * 3 + 2]
+        indices[i] = i
+      }
+      // Create the normals buffer for the mesh
+      val normals = FloatArray(positions.size) { 0.0f }
+      // Create the colors buffer for the mesh
+      val color =
+          when (roomFace.label) {
+            MRUKLabel.FLOOR -> Color.valueOf(0.2f, 0.6f, 0.2f, 1.0f) // Green
+            MRUKLabel.CEILING -> Color.valueOf(0.8f, 0.8f, 0.8f, 1.0f) // White
+            MRUKLabel.WALL_FACE -> Color.valueOf(0.6f, 0.6f, 0.8f, 1.0f) // Blue
+            MRUKLabel.INVISIBLE_WALL_FACE -> Color.valueOf(0.8f, 0.3f, 0.8f, 1.0f) // Purple
+            MRUKLabel.INNER_WALL_FACE -> Color.valueOf(0.4f, 0.4f, 0.6f, 1.0f) // Dark Blue
+            MRUKLabel.WINDOW_FRAME -> Color.valueOf(0.7f, 0.9f, 1.0f, 1.0f) // Light Blue
+            MRUKLabel.DOOR_FRAME -> Color.valueOf(0.6f, 0.4f, 0.2f, 1.0f) // Brown
+            else -> Color.valueOf(1.0f, 1.0f, 1.0f)
+          }
+      val colors: IntArray = IntArray(positions.size / 3) { color.toArgb() }
+      // Create the UV buffer for the mesh
+      val uvs: FloatArray = FloatArray(positions.size / 3 * 2)
+
+      // Create the mesh
+      val sceneMesh =
+          SceneMesh.meshWithMaterials(
+              positions,
+              normals,
+              uvs,
+              colors,
+              indices,
+              intArrayOf(0, indices.size),
+              arrayOf(SceneMaterial(SceneTexture(Color.valueOf(1.0f, 1.0f, 1.0f, 1.0f)))),
+              false,
+          )
+      sceneMesh
+    }
+  }
+
+  private fun getSelectedSceneModel(sceneModelSpinner: Spinner): SceneModel {
+    return sceneModelSpinner.selectedItem?.let { item ->
+      when (item as? String) {
+        "Scene V1 (Basic Scene)" -> SceneModel.V1
+        "Scene V2 (High-Fidelity Scene)" -> SceneModel.V2
+        "Scene V2 with Fallback to Scene V1" -> SceneModel.V2_FALLBACK_V1
+        else -> SceneModel.V1
+      }
+    } ?: SceneModel.V1 // Provide a default value if null
+  }
+
   companion object {
     const val TAG: String = "MrukAnchorMeshSample"
     const val PERMISSION_USE_SCENE: String = "com.oculus.permission.USE_SCENE"
     const val REQUEST_CODE_PERMISSION_USE_SCENE: Int = 1
     const val REQUEST_CODE_PERMISSION_WRITE_EXTERNAL_STORAGE = 2
+    const val ROOM_MESH_ID: String = "mesh://roommesh"
   }
 }